+++ packages/apps/kadena-docs/src/components/Layout/components/Aside/index.ts
@@ -1,0 +1,2 @@
+export * from './AsideLink';
+export * from './AsideStyles';
