+++ packages/apps/kadena-docs/src/components/Layout/components/index.ts
@@ -3,5 +3,6 @@ export * from './Footer';
 export * from './DocsLogo';
 export * from './Main';
 export * from './styles';
-export * from './AsideStyles';
+export * from './Aside';
+export * from './ArticleStyles';
 export * from './TitleHeader';
